backend-ai/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
Pillow==11.3.0
jieba==0.42.1
rapidocr-onnxruntime==1.4.4
